refactor!: fhe_http_python for package build - Create a folder named 'fhe_http', which matches the package name, and move assembler into the folder as a module. - Update import package in tests script
