+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
 
-members = ["fhe_http_core", "fhe_http_python/fhe"]
+members = ["fhe_http_core", "fhe_http_python"]
