+++ fhe_http_python/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "fhe"
+name = "fhe_http"
 version = "0.1.0"
 edition = "2021"
 
@@ -10,7 +10,7 @@ name = "fhe"
 
 [dependencies]
 pyo3 = { version = "0.22.1", features = ["extension-module"] }
-fhe_http_core = { path = "../../fhe_http_core" }
+fhe_http_core = { path = "../fhe_http_core" }
 project-root = "0.2.2"
 serde_json = "1.0.116"
 zstd = "0.13.1"
@@ -18,4 +18,4 @@ concat-idents = "=1.1.5"
 
 
 [package.metadata.maturin]
-name = "fhe"
+name = "fhe_http"
