--- fhe_http_python/fhe/pyproject.toml
@@ -1,15 +1,0 @@
-[build-system]
-requires = ["maturin>=1.5,<2.0"]
-build-backend = "maturin"
-
-[project]
-name = "fhe"
-requires-python = ">=3.8"
-classifiers = [
-    "Programming Language :: Rust",
-    "Programming Language :: Python :: Implementation :: CPython",
-    "Programming Language :: Python :: Implementation :: PyPy",
-]
-dynamic = ["version"]
-[tool.maturin]
-features = ["pyo3/extension-module"]
