+++ fhe_http_python/fhe_http/__init__.py
@@ -1,0 +1,5 @@
+from .fhe import *
+from .fhe import __all__
+from .assembler.assembler import Assembler
+
+__all__ = __all__ + ["Assembler"]
