+++ fhe_http_python/pyproject.toml
@@ -1,11 +1,16 @@
 [build-system]
-requires = [
-    "setuptools>=68.2.2",
-    "wheel>=0.43.0",
-    "setuptools-rust>=1.9.0",
-    "ply>=3.11",
-    "maturin>=1.5.0",
-    "semantic-version>=2.10.0",
-    "tomli>=2.0.1",
+requires = ["maturin>=1.5,<2.0"]
+build-backend = "maturin"
+
+[project]
+name = "fhe-http"
+requires-python = ">=3.8"
+classifiers = [
+    "Programming Language :: Rust",
+    "Programming Language :: Python :: Implementation :: CPython",
+    "Programming Language :: Python :: Implementation :: PyPy",
 ]
-build-backend = "setuptools.build_meta"
+dynamic = ["version"]
+
+[tool.maturin]
+features = ["pyo3/extension-module"]
